+++ src/terminal/mod.rs
@@ -120,7 +120,7 @@ impl TerminalCapabilities {
                 TerminalCapabilities {
                     name: "iTerm2".to_string(),
                     style: StyleCapability::Ansi(AnsiStyle),
-                    links: LinkCapability::OSC8(self::osc::OSC8Links),
+                    links: LinkCapability::OSC8(self::osc::OSC8Links::for_localhost()),
                     image: ImageCapability::ITerm2(self::iterm2::ITerm2Images),
                     marks: MarkCapability::ITerm2(self::iterm2::ITerm2Marks),
                 }
@@ -131,7 +131,7 @@ impl TerminalCapabilities {
                 TerminalCapabilities {
                     name: "Terminology".to_string(),
                     style: StyleCapability::Ansi(AnsiStyle),
-                    links: LinkCapability::OSC8(self::osc::OSC8Links),
+                    links: LinkCapability::OSC8(self::osc::OSC8Links::for_localhost()),
                     image: ImageCapability::Terminology(self::terminology::TerminologyImages),
                     marks: MarkCapability::None,
                 }
@@ -142,7 +142,7 @@ impl TerminalCapabilities {
                 TerminalCapabilities {
                     name: "VTE 50".to_string(),
                     style: StyleCapability::Ansi(AnsiStyle),
-                    links: LinkCapability::OSC8(self::osc::OSC8Links),
+                    links: LinkCapability::OSC8(self::osc::OSC8Links::for_localhost()),
                     image: ImageCapability::None,
                     marks: MarkCapability::None,
                 }